feat: add ModuleConfigurationDialogComponent for module configuration management
